% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{download_font}
\alias{download_font}
\title{Download font files}
\usage{
download_font(id, output_dir, ...)
}
\arguments{
\item{id}{Id of the font, correspond to column \code{id} from \code{\link{get_all_fonts}}.}

\item{output_dir}{Output directory where to save font files.}

\item{...}{Additional parameters to API query, for example \code{variant = "regular"}.}
}
\value{
a character vector of the filepaths extracted to, invisibly.
}
\description{
Download font files
}
\examples{
\dontrun{

download_font("robot", "path/to/directory")

}
}
