% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_all_fonts}
\alias{get_all_fonts}
\title{Get infos about all fonts available}
\usage{
get_all_fonts()
}
\value{
a \code{data.frame}.
}
\description{
Get infos about all fonts available
}
\examples{
\dontrun{
# Retrieve all fonts currently available
all_fonts <- get_all_fonts()

}
}
