% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_font_info}
\alias{get_font_info}
\title{Get detailled infos about one font}
\usage{
get_font_info(id, subsets = NULL)
}
\arguments{
\item{id}{Id of the font, correspond to column \code{id} from \code{\link{get_all_fonts}}.}

\item{subsets}{Select charsets, for example \code{"latin"}.}
}
\value{
a \code{data.frame}.
}
\description{
Get detailled infos about one font
}
\examples{
\donttest{

roboto <- get_font_info("roboto")

}
}
