% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{included_fonts}
\alias{included_fonts}
\title{Detail about included fonts.}
\format{
A data frame with 8 rows and 5 variables:
\describe{
  \item{id}{Id for the font.}
  \item{family}{Name of the font.}
  \item{category}{Category.}
  \item{version}{Version number.}
  \item{lastModified}{Last modified date.}
}
}
\source{
\url{https://google-webfonts-helper.herokuapp.com/fonts}
}
\usage{
included_fonts
}
\description{
Id and version of fonts included and available through \code{\link{use_pkg_gfont}}.
}
\keyword{datasets}
