% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_font.R
\name{setup_font}
\alias{setup_font}
\title{Setup a font to be used in Shiny or Markdown}
\usage{
setup_font(id, output_dir, variants = NULL)
}
\arguments{
\item{id}{Id of the font, correspond to column \code{id} from \code{\link{get_all_fonts}}.}

\item{output_dir}{Output directory where to save font and CSS files. Must be a directory.}

\item{variants}{Variant(s) to download, default is to includes all available ones.}
}
\value{
To directories will be created (if they do not exist): \strong{fonts} and \strong{css}.
}
\description{
Setup a font to be used in Shiny or Markdown
}
\examples{
\donttest{
setup_font(
  id = "open-sans-condensed",
  output_dir = "path/to/www"
)
}
}
