% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tag_example.R
\name{tag_example}
\alias{tag_example}
\title{Generate HTML tags used in examples}
\usage{
tag_example(class = NULL)
}
\arguments{
\item{class}{Class of the main div.}
}
\value{
HTML tags.
}
\description{
Generate HTML tags used in examples
}
\examples{
tag_example()
}
