% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/css.R
\name{use_gfont}
\alias{use_gfont}
\title{Use a Google Font}
\usage{
use_gfont(font = c("roboto", "open-sans", "lato", "montserrat"))
}
\arguments{
\item{font}{Name of the font}
}
\value{
an \code{htmlDependency}.
}
\description{
Use a Google Font
}
\examples{
# todo
}
