% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{download_font}
\alias{download_font}
\title{Download font files}
\usage{
download_font(id, output_dir, variants = NULL, ...)
}
\arguments{
\item{id}{Id of the font, correspond to column \code{id} from \code{\link{get_all_fonts}}.}

\item{output_dir}{Output directory where to save font files.}

\item{variants}{Variant(s) to download, default is to includes all available ones.}

\item{...}{Additional parameters to API query.}
}
\value{
a character vector of the filepaths extracted to, invisibly.
}
\description{
Download font files
}
\examples{
\dontrun{

# Download Roboto font
download_font(
  id = "roboto",
  output_dir = "path/to/directory"
)

# Get only regular, italic and bold
download_font(
  id = "roboto",
  output_dir = "path/to/directory",
  variants = c("regular", "300italic", "700")
)

}
}
