% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_all_fonts}
\alias{get_all_fonts}
\title{Get infos about all fonts available}
\usage{
get_all_fonts(...)
}
\arguments{
\item{...}{Arguments passed to \code{crul::HttpClient$new}.}
}
\value{
a \code{data.frame}.
}
\description{
Retrieve from API all fonts currently available.
 Use the \code{id} field in other functions to reference the font you want to use.
}
\examples{
if (interactive()) {

 # Retrieve all fonts currently available
 all_fonts <- get_all_fonts()

}
}
