% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gfonts-package.R
\docType{package}
\name{gfonts}
\alias{gfonts}
\title{Use Google fonts offline}
\description{
Download Google fonts and generate CSS to use in {rmarkdown} documents and
{shiny} applications. Some popular fonts are included and ready to use.
}
\author{
Victor Perrier (@dreamRs_fr)
}
