% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_font.R
\name{use_font}
\alias{use_font}
\title{Use a downloaded font in Shiny or Markdown}
\usage{
use_font(id, css_path, selector = "body", css = NULL)
}
\arguments{
\item{id}{Id of the font downloaded.}

\item{css_path}{Path to CSS generated by \code{\link{setup_font}}.}

\item{selector}{CSS selector for which to use the font,
usually an HTML tag, default to \code{"body"} (all document).}

\item{css}{CSS variables needed to use font, normally this should be automatic.}
}
\value{
an HTML tag with a dependency.
}
\description{
Use a downloaded font in Shiny or Markdown
}
